third_party/blink/web_tests/platform/generic/accessibility/ellipsis-text-expected.txt

text
My Writing Nook for iPad. More text, more text, more text.
My Writing Nook for iPad. More text, more text, more text.
text

This test makes sure that the bounds of text that overflows with ellipsis is correct (shortened to the ellipsis that is).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS clippedWidth < fullWidth is true
PASS successfullyParsed is true

TEST COMPLETE
